requirements.txt
jupyter==1.1.1
notebook==7.4.3
jupyterlab==4.4.3
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0 
gspread>=6.2.1
oauth2client>=4.1.3
ruamel.yaml>=0.18.6
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
gspread>=5.0.0
pandas>=1.3.0